App that create, retrieve, update & delete ideas
This app was created as to make a REST API, REST is an Abbreviation for "Representative State Transfer". This kind of API's are highly used now adays as SOAP wasn't useful as required. This App creation includes knowledge about Server and Client, URI and its meaning. Used VS code and Postman as per required and result it here!.